validate token claims on socket connection
